Forgot to migrate and fixed error messages not showing up on register page sometimes
Error messages not showing up had to do with password fields. If everything else was correct and password fields didn't match, a validation error was raised in a FORM function. However, in views we are using messages and in the template we used messages - therefore the validation error was never displayed. Because of this error, it meant that the form was "not valid" and none of the other error messages would show either since we check for those errors inside of the form.is_valid() if statement.